rename resources
