+++ service/controller/resource/monitoring/remotewriteapiendpointconfigsecret/doc.go
@@ -1,4 +1,4 @@
-package remotewriteagentconfigsecret
+package remotewriteapiendpointsecret
 
 // This packages wraps the remote write config in a secret.
 // This is to be used by prometheus operator to configure prometheus.
