+++ service/controller/resource/monitoring/remotewriteapiendpointconfigsecret/test/case-1-awsconfig.golden
