+++ service/controller/resource/monitoring/remotewriteapiendpointconfigsecret/test/case-2-azureconfig.golden
