+++ service/controller/resource/monitoring/remotewriteapiendpointconfigsecret/test/case-4-control-plane.golden
