+++ service/controller/resource/monitoring/remotewriteapiendpointsecret/resource_test.go
@@ -1,4 +1,4 @@
-package remotewriteagentsecret
+package remotewriteapiendpointsecret
 
 import (
 	"context"
