+++ service/controller/resource/monitoring/remotewriteapiendpointsecret/test/case-2-azureconfig.golden
