+++ service/controller/resource/monitoring/remotewriteapiendpointsecret/test/case-3-kvmconfig.golden
