+++ service/controller/resource/monitoring/remotewriteapiendpointsecret/test/case-4-control-plane.golden
