+++ service/controller/resource/monitoring/remotewriteapiendpointsecret/test/case-5-cluster-api-v1alpha3.golden
